Add Suspense into page
